LayoutTests/inspector/console/console-object-constructor-name.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/console-test.js"></script>
<script>

function Parent() { }
function Child() { }
Child.prototype = new Parent();
var outer = { inner: function() { } };
console.log(new Parent());
console.log(new Child());
console.log(new outer.inner());

function test()
{
    InspectorTest.expandConsoleMessages();
    InspectorTest.dumpConsoleMessages();
    InspectorTest.completeTest();
}

</script>
</head>

<body onload="runTest()">
<p>
Tests that the name of the function invoked as object constructor will be displayed as its type in the front-end.
<a href="https://bugs.webkit.org/show_bug.cgi?id=50063">Bug 50063.</a>
</p>

</body>
</html>
